+++ HISTORY.rst
@@ -96,6 +96,7 @@ History
     * Change the query to compute total_orders and total_leadorders (rudaporto).
     * Review list of fields to be serialized in Professional, Assignment, Order and Project (rudaporto).
     * Fix all to_dict methods so now they can be cached without pickle errors (rudaporto).
+    * Card #582: fix assignment slug creation based on the order slug (rudaporto).
 
 2.1.36 (2017-08-06)
 -------------------